requirements.txt
numpy
pandas
scipy
scikit-learn
javabridge
python-weka-wrapper3